Start moving user facing logic into TelegramClient
