+++ example/main.dart
@@ -1,11 +1,7 @@
 import '../telegram/client/telegram_client.dart';
 
-class A {
-  static int b = 5;
-}
-
 main() async {
-  var client = TelegramClient("aa", 1, "a");
+  var client = TelegramClient("aa", 1, "a", {});
   await client.connect();
   print("Connected successfully");
 }
